refactor(domain): LessonReport 도메인 분리
domains/lesson 하위에 있던 LessonReport 관련 파일을 별도 도메인
domains/lessonReport 로 분리. 향후 리뷰/콜백/파일 첨부 등 리포트
플로우 확장 시 lesson 도메인 책임과 분리해 관리하기 위함.

- 신규: domains/lessonReport/{domain,repository,adaptor,exception}
- 별도 LessonReportErrorCode (prefix: LESSON_REPORT_001~006)
- 기존 LessonErrorCode/LessonExceptions에서 Report 항목 제거
- LESSON_UNAUTHORIZED_ACCESS 코드 LESSON_010 → LESSON_004 재번호
