+++ SClass-Domain/src/main/kotlin/com/sclass/domain/domains/lessonReport/domain/LessonReport.kt
@@ -1,4 +1,4 @@
-package com.sclass.domain.domains.lesson.domain
+package com.sclass.domain.domains.lessonReport.domain
 
 import com.sclass.domain.common.model.BaseTimeEntity
 import jakarta.persistence.Column
@@ -31,35 +31,25 @@ class LessonReport(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
-
     @Column(name = "lesson_id", nullable = false)
     val lessonId: Long,
-
     @Column(nullable = false)
     val version: Int,
-
     @Column(name = "submitted_by_user_id", nullable = false, length = 26)
     val submittedByUserId: String,
-
     @Column(columnDefinition = "TEXT", nullable = false)
     var content: String,
-
     @Column(name = "report_file_id", length = 26)
     var reportFileId: String? = null,
-
     @Column(name = "class_video_file_id", length = 26)
     var classVideoFileId: String? = null,
-
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
     var status: LessonReportStatus = LessonReportStatus.PENDING_REVIEW,
-
     @Column(name = "reviewed_by_user_id", length = 26)
     var reviewedByUserId: String? = null,
-
     @Column(name = "reviewed_at")
     var reviewedAt: LocalDateTime? = null,
-
     @Column(name = "reject_reason", length = 1000)
     var rejectReason: String? = null,
 ) : BaseTimeEntity() {
