+++ SClass-Domain/src/main/kotlin/com/sclass/domain/domains/lessonReport/domain/LessonReportStatus.kt
@@ -1,4 +1,4 @@
-package com.sclass.domain.domains.lesson.domain
+package com.sclass.domain.domains.lessonReport.domain
 
 enum class LessonReportStatus {
     PENDING_REVIEW,
